requirements.txt
flask==3.0.0
flask-cors==4.0.0
alpaca-py==0.21.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
gunicorn==21.2.0
